requirements.txt
azure-ai-projects>=2.0.0
azure-identity>=1.17.0
pytest>=8.0.0
